% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_group_exp}
\alias{create_group_exp}
\title{This function focuses on cleaning scRNA-seq and SRT datasets.}
\usage{
create_group_exp(sc_exp, sc_label)
}
\arguments{
\item{sc_exp}{scRNA-seq matrix, genes * cells. The format should be raw-counts. The matrix need include gene names and cell names.}

\item{sc_label}{cell type information. The cells are need be divided into multiple category.}
}
\description{
This function focuses on cleaning scRNA-seq and SRT datasets.
}
\examples{
data("PDAC")
refmu = create_group_exp(sc_exp,sc_label)
}
