% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_process}
\alias{data_process}
\title{Data Processing for Spatial and Single-cell Transcriptomics.
Processes and cleans spatial transcriptomics (SRT) and single-cell RNA-seq (scRNA-seq)
data by filtering genes and cells/spots based on specified criteria. It aligns the
gene expression data between scRNA-seq and SRT datasets by retaining only the common genes.}
\usage{
data_process(
  sc_exp,
  sc_label,
  spot_exp,
  spot_loc,
  gene_det_in_min_cells_per = 0.01,
  expression_threshold = 0,
  nUMI = 100,
  verbose = FALSE,
  depthscale = 1e+06,
  clean.only = TRUE
)
}
\arguments{
\item{sc_exp}{A matrix of scRNA-seq data (genes x cells) containing expression counts.}

\item{sc_label}{A vector of labels corresponding to the cells in the sc_exp matrix.}

\item{spot_exp}{A matrix of spatial transcriptomics data (genes x spots) containing expression counts.}

\item{spot_loc}{A matrix of coordinates for each spot in the spot_exp matrix.}

\item{gene_det_in_min_cells_per}{Minimum percentage of cells in which a gene must be detected.}

\item{expression_threshold}{The minimum expression level a gene must exhibit to be considered expressed.}

\item{nUMI}{Minimum number of UMIs (unique molecular identifiers) required for a cell/spot to be included.}

\item{verbose}{Logical; if TRUE, detailed processing information will be printed.}

\item{depthscale}{A scaling factor used for normalization (default is 100,000).}

\item{clean.only}{Logical; if TRUE, only performs data cleaning without additional normalization.}
}
\value{
A list containing filtered and processed scRNA-seq and SRT data:
  - `sc_exp`: Filtered scRNA-seq expression matrix.
  - `sc_label`: Corresponding cell labels for the filtered scRNA-seq data.
  - `spot_exp`: Filtered SRT expression matrix.

#data(Fishplus)
#datax = data_process(sc_exp = sc_exp, sc_label = sc_label, spot_exp = spot_exp, spot_loc = spot_loc)
}
\description{
Data Processing for Spatial and Single-cell Transcriptomics.
Processes and cleans spatial transcriptomics (SRT) and single-cell RNA-seq (scRNA-seq)
data by filtering genes and cells/spots based on specified criteria. It aligns the
gene expression data between scRNA-seq and SRT datasets by retaining only the common genes.
}
