% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dis_weight}
\alias{dis_weight}
\title{Calculate Weighted Adjacency Matrix for Spatial Data}
\usage{
dis_weight(
  spot_loc,
  spot_exp,
  k = 6,
  quantile_prob_bandwidth = 1/3,
  method = "Hex",
  coord_type = "grid"
)
}
\arguments{
\item{spot_loc}{A matrix of coordinates for each spot, typically named (x, y).}

\item{spot_exp}{A matrix of spatial transcriptomics data (genes x spots) containing expression counts.}

\item{k}{Integer; number of neighboring spots to consider when constructing the spatial neighboring graph.}

\item{quantile_prob_bandwidth}{The bandwidth quantile for weighting the adjacency matrix.}

\item{method}{Method used for constructing the spatial neighboring graph. Currently supports only 'Hex'.}

\item{coord_type}{Type of coordinates system used, typically 'grid' for standard spatial transcriptomics data.}
}
\value{
A list containing:
  - `dis_weight`: The Gaussian-weighted adjacency matrix based on spatial distances and connectivity.
  - `weight_adj`: The unweighted adjacency matrix derived from spatial connectivity alone.
}
\description{
This function computes the weighted adjacency matrix for spatial data based on the
connectivity and distances between spots. It supports only the 'Hex' method for
defining spatial neighbors in a hexagonal grid system.
}
\examples{
# Assuming `coordinates` is a matrix of spatial coordinates and `expression` is the corresponding expression data
data(Fishplus)
weights <- dis_weight(spot_loc = spot_loc, spot_exp = spot_exp, k = 6, method = "Hex", coord_type = "grid")
}
