% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fishplus}
\alias{Fishplus}
\alias{sc_exp}
\title{Input data of the STged}
\format{
a list of matrices contains SRT gene expression matrix with coordinates information and corresponding cell type proportion information,
and annotated scRNA-seq data as cell type reference (bottom).
}
\usage{
data(Fishplus)
}
\description{
Input data of the STged
}
\examples{
data(Fishplus)
}
\keyword{dataset}
