% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MUR.STged}
\alias{MUR.STged}
\title{Multiplicative update rule (MUR) for Spatial Transcriptomics Gene Expression Deconvolution (MUR.STged)
This function performs the gene expression deconvolution using a multilevel perceptron approach for
spatial transcriptomics data.}
\usage{
MUR.STged(
  srt_exp = srt_exp,
  ref_exp = ref_exp,
  beta.type = beta.type,
  W = W,
  lambda1 = lambda1,
  lambda2 = lambda2,
  tau1 = 0.1,
  tau2 = 0.1,
  cutoff = 0.05,
  epsilon = 1e-05,
  maxiter = 100
)
}
\arguments{
\item{srt_exp}{A matrix of spatial transcriptomics RNA-seq gene expression data (genes x spots), containing raw counts.}

\item{ref_exp}{A matrix of reference single-cell RNA-seq gene expression data (genes x cell types).}

\item{beta.type}{A matrix describing the initial cell type proportions at each spot (spots x cell types).}

\item{W}{A spatial weight matrix (spots x spots) describing the spatial correlation between spots.}

\item{lambda1}{Regularization parameter for the graph regularization term, controlling spatial smoothness across neighboring spots. Automatically determined if set to NULL.}

\item{lambda2}{Regularization parameter for the prior regularization term, aligning estimated gene expression profiles with biologically consistent patterns. Automatically determined if set to NULL.}

\item{tau1}{Scaling factor for lambda1, adjusting the strength of spatial regularization. Defaults to 0.1.}

\item{tau2}{Scaling factor for lambda2, adjusting alignment with prior biological information. Defaults to 0.1.}

\item{cutoff}{Cutoff value for cell type proportion to filter insignificant cell type contributions. Default is 0.05.}

\item{epsilon}{Convergence threshold for stopping the algorithm. Default is 1e-5.}

\item{maxiter}{Maximum number of iterations allowed in the algorithm. Default is 100.}
}
\value{
A list containing the following elements:
  \item{V.hat}{Estimated gene expression matrix (genes x spots).}
  \item{F_list}{List of estimated cell type-specific gene expression matrices.}
  \item{lambda1}{Selected value of lambda1.}
  \item{lambda2}{Selected value of lambda2.}
  \item{beta}{Final estimated cell type proportions matrix (spots x cell types).}
  \item{obj.loss}{Objective loss value of the final model.}
}
\description{
Multiplicative update rule (MUR) for Spatial Transcriptomics Gene Expression Deconvolution (MUR.STged)
This function performs the gene expression deconvolution using a multilevel perceptron approach for
spatial transcriptomics data.
}
\examples{
\dontrun{
# Example usage:
result <- MUR.STged(srt_exp = spatial_exp, ref_exp = reference_exp, beta.type = beta,
                    W = spatial_weights, lambda1 = NULL, lambda2 = NULL,
                    tau1 =0.1, tau2 =0.1,cutoff = 0.05,
                    epsilon = 1e-5, maxiter = 100)
}
}
