% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{STged}
\alias{STged}
\title{Spatial Transcriptomics Gene Expression Deconvolution (STged)}
\usage{
model_results <- STged(sc_exp = sc_exp, sc_label = sc_label,
                      spot_exp = spot_exp, spot_loc = spot_loc, beta = beta)
}
\arguments{
\item{sc_exp}{A matrix of single-cell RNA-seq gene expression data (genes x cells), containing raw counts.}

\item{sc_label}{A vector or factor of cell type labels corresponding to each cell in the sc_exp data.}

\item{spot_exp}{A matrix of spatial transcriptomics RNA-seq gene expression data (genes x spots), containing raw counts.}

\item{spot_loc}{A matrix specifying the coordinates of each spot (spots x coordinates), typically named (x, y).}

\item{beta}{A matrix describing the cell type proportions at each spot (spots x cell types).}

\item{gene_det_in_min_cells_per}{Minimum percentage of genes that must be detected in a cell, used as a filtering criterion.}

\item{expression_threshold}{Minimum expression level to consider a gene as expressed, used for data cleaning.}

\item{nUMI}{Minimum number of read counts required to consider a cell or spot, used for data cleaning.}

\item{verbose}{Logical; if TRUE, prints details of the data processing steps.}

\item{clean.only}{Logical; if TRUE, applies only data cleaning steps without normalization.}

\item{depthscale}{A scaling factor used for normalization (default is 100,000).}

\item{python_env}{Path to the Python environment, used for calling Python functions.}

\item{truncate}{Logical; if TRUE, truncates gene expression data to reduce the impact of extreme values.}

\item{qt}{Quantile used for Winsorizing expression values to control the effects of outliers.}

\item{knei}{Integer; number of neighboring spots to consider when constructing the spatial neighboring graph.}

\item{methodL}{Method used for constructing the spatial neighboring graph, details provided by the Squidpy package.}

\item{coord_type}{Type of coordinates system used, e.g., 'grid' as described by Squidpy.}

\item{quantile_prob_bandwidth}{Bandwidth for the spatial kernel used in constructing the neighboring graph.}

\item{lambda1}{Regularization parameter for the graph regularization term, controlling spatial smoothness across neighboring spots. Automatically determined if set to NULL.}

\item{lambda2}{Regularization parameter for the prior regularization term, aligning estimated gene expression profiles with biologically consistent patterns. Automatically determined if set to NULL.}

\item{tau2}{Scaling factor for lambda2, adjusting alignment with prior biological information. Defaults to 0.1.}

\item{cutoff}{Cutoff value for cell type proportion to filter insignificant cell type contributions.}

\item{maxiter}{Maximum number of iterations allowed in the algorithm.}

\item{epsilon}{Convergence threshold for stopping the algorithm.}

\item{rho}{Penalty parameter in the ADMM algorithm, with adaptive adjustments.}

\item{rho.incr}{Increment step for varying the penalty parameter rho.}

\item{rho.max}{Maximum allowable value for rho.}
}
\value{
A list containing model estimation results, including estimated parameters and diagnostics.
}
\description{
Spatial Transcriptomics Gene Expression Deconvolution (STged)
}
