% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update.Fk}
\alias{update.Fk}
\title{Internal helper function for updating Fk}
\usage{
\method{update}{Fk}(
  Fk_old = Fk_old,
  srt_exp = srt_exp,
  Rk = Rk,
  Ak = Ak,
  Bk = Bk,
  Wk = Wk,
  Dk = Dk,
  lambda1 = lambda1,
  lambda2 = lambda2
)
}
\description{
Internal helper function for updating Fk
}
\keyword{internal}
