% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_group_exp}
\alias{create_group_exp}
\title{Create Grouped Expression Matrix by Cell Type}
\usage{
create_group_exp(sc_exp, sc_label)
}
\arguments{
\item{sc_exp}{A matrix of single-cell RNA-seq data (genes x cells), containing expression counts.}

\item{sc_label}{A vector of labels corresponding to the cells in the sc_exp matrix.}
}
\value{
A matrix of mean expression values for each cell type (genes x cell types).
# Assuming `sc_data` is a matrix of gene expression data and `cell_types` is a vector of cell type labels
}
\description{
This function computes the mean expression for each cell type from scRNA-seq data.
It groups cells by their specified type and calculates the mean of all cells within
each type to create a summarized expression profile per cell type.
}
\examples{
data(Fishplus)
refmu = create_group_exp(sc_exp,sc_label)
}
