% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dis_weight}
\alias{dis_weight}
\title{This function focuses on cleaning scRNA-seq and SRT datasets.}
\usage{
dis_weight(
  spot_loc = spot_loc,
  spot_exp = spot_exp,
  k = 6,
  quantile_prob_bandwidth = 1/3,
  method = "Hex",
  coord_type = "grid"
)
}
\arguments{
\item{spot_loc}{coordinate matrix, spots * coordinates. The matrix need include spot names and coordinate name (x, y).}

\item{spot_exp}{stRNA-seq matrix, genes * spots. The format should be raw-counts. The matrix need include gene names and spot names.}

\item{k}{number of neighbor spots for construct spatial neighboring graph, details refer to Squidpy.}

\item{quantile_prob_bandwidth}{selection of bandwidth of the spatial kernel of each spot.}

\item{method}{the method used to construct spatial neighboring graph, details refer to Squidpy.}

\item{coord_type}{grid coordinates, details refer to Squidpy.}
}
\description{
This function focuses on cleaning scRNA-seq and SRT datasets.
}
