% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{STged}
\alias{STged}
\title{STged: Gene expression deconvolution for spatial transcriptomic data}
\usage{
model.est = STged(sc_exp = sc_exp, sc_label = sc_label, spot_exp = spot_exp, spot_loc = spot_loc)
}
\arguments{
\item{sc_exp}{scRNA-seq matrix, genes * cells. The format should be raw-counts. The matrix need include gene names and cell names.}

\item{sc_label}{cell type information. The cell need be divided into multiple categories.}

\item{spot_exp}{stRNA-seq matrix, genes * spots. The format should be raw-counts. The matrix need include gene names and spot names.}

\item{spot_loc}{coordinate matrix, spots * coordinates. The matrix need include spot names and coordinate name (x, y).}

\item{beta}{cell type proportion matrix, spots * cell types.}

\item{gene_det_in_min_cells_per}{a floor variable. minimum percent # of genes that need to be detected in a cell.}

\item{expression_threshold}{a floor variable. Threshold to consider a gene expressed.}

\item{nUMI}{a floor variable.     minimum # of read count that need to be detected in a cell or spot.}

\item{verbose}{a logical variable that defines whether to print the processing flow of data process.}

\item{clean.only}{a logical variable that defines whether to normalize data with log transform.}

\item{python_env}{the path of python environment.}

\item{truncate}{a logical variable. that defines whether to truncate the gene expression data for both scRNA-seq and SRT.}

\item{qt}{Winsorize expression values to prevent outliers. Values below this quantile and above 1-this quantile will be set to the quantile value.}

\item{knei}{number of neighbor spots for construct spatial neighboring graph, details refer to Squidpy.}

\item{methodL}{the method used to construct spatial neighboring graph, details refer to Squidpy.}

\item{coord_type}{grid coordinates, details refer to Squidpy.}

\item{quantile_prob_bandwidth}{selection of bandwidth of the spatial kernel of each spot.}

\item{lambda1}{tuning parameter to balance the graph regularization term. If the tuning parameter is set to NULL, then, we will adopt the value in our algorithm.}

\item{lambda2}{tuning parameter to balance the prior regularization term. If the tuning parameter is set to NULL, then, we will adopt the value in our algorithm.}

\item{cutoff}{a cutoff value of cell type proportion.}

\item{rho}{the penalty parameter in the ADMM algorithm.}

\item{rho.incr}{the increase step parameter for varying penalty parameter rho.}

\item{rho.max}{the maximum value of rho.}

\item{maxiter}{a positive integer represents the maximum number of updating algorithm. Default setting is 100.}

\item{epsilon}{a parameter represents the stop criterion.}
}
\description{
STged is a gene expression deconvolution mathod. STged employs graph-guided spatial correlations and prior-guided gene expression similarities within a non-negative least-squares regression framework.
}
